requirements.txt
# 语音输入助手项目依赖
# 语音识别核心库
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# 音频处理
pyaudio==0.2.11
sounddevice==0.4.6
librosa==0.10.1

# GUI框架
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# 系统操作
pyautogui==0.9.54
pynput==1.7.6
pywin32==306
psutil==5.9.6

# 网络请求和API
requests==2.31.0
openai==1.3.5

# 配置文件处理
pyyaml==6.0.1
configparser==6.0.0

# 日志和工具
loguru==0.7.2
click==8.1.7

# 可选：本地大模型支持
transformers==4.35.2
sentence-transformers==2.2.2 